+++ aquadoggo/src/test_utils/mod.rs
@@ -7,12 +7,12 @@ pub mod helpers;
 mod node;
 mod runner;
 
-pub use client::{graphql_test_client, TestClient};
+pub use client::{http_test_client, TestClient};
 pub use config::TestConfiguration;
 pub use db::{drop_database, initialize_db, initialize_sqlite_db};
 pub use helpers::{build_document, doggo_fields, doggo_schema, schema_from_fields};
 pub use node::{
     add_blob, add_document, add_schema, add_schema_and_documents, assert_query,
-    populate_and_materialize, populate_store_config, update_document, TestNode,
+    populate_and_materialize, populate_store_config, update_blob, update_document, TestNode,
 };
 pub use runner::{test_runner, test_runner_with_manager, TestNodeManager};