+++ aquadoggo/src/test_utils/runner.rs
@@ -8,7 +8,7 @@ use p2panda_rs::identity::KeyPair;
 use tokio::runtime::Builder;
 use tokio::sync::Mutex;
 
-use crate::config::{BLOBS_DIR_NAME, BLOBS_SYMLINK_DIR_NAME};
+use crate::config::BLOBS_DIR_NAME;
 use crate::context::Context;
 use crate::db::Pool;
 use crate::db::SqlStore;
@@ -102,12 +102,12 @@ pub fn test_runner<F: AsyncTestFn + Send + Sync + 'static>(test: F) {
         let (_config, pool) = initialize_db().await;
         let store = SqlStore::new(pool);
 
-        // Construct tempfile directory for the test runner.
+        // Construct temporary directory for the test runner
         let tmp_dir = tempfile::TempDir::new().unwrap();
         let blob_dir_path = tmp_dir.path().join(BLOBS_DIR_NAME);
-        fs::create_dir_all(blob_dir_path.join(BLOBS_SYMLINK_DIR_NAME)).unwrap();
+        fs::create_dir_all(&blob_dir_path).unwrap();
 
-        // Construct node config supporting any schema.
+        // Construct node config supporting any schema
         let mut cfg = Configuration::default();
         cfg.blob_dir = Some(blob_dir_path);